refactor(store/state): factor out account-related methods
